requirements.txt

# pip install -r requirements.txt

asgiref
Django
psycopg2
python-dotenv
sqlparse
tzdata
whitenoise
djangorestframework
django-cors-headers
pyjwt
pillow
channels
daphne
docker
mongoengine
django_mongoengine
django-rest-framework-mongoengine
redis
celery
channels_redis
Twisted[http2,tls]
django-environ
logtail-python
